Moved some longer running tests to IT tests
And allow longer runtime for slower CPUs
