+++ src/test/java/org/jsoup/parser/HtmlParserTest.java
@@ -992,29 +992,6 @@ public class HtmlParserTest {
         assertTrue(System.currentTimeMillis() - start < 1000);
     }
 
-    @Test public void handlesDeepStack() {
-        // inspired by http://sv.stargate.wikia.com/wiki/M2J and https://github.com/jhy/jsoup/issues/955
-        // I didn't put it in the integration tests, because explorer and intellij kept dieing trying to preview/index it
-
-        // Arrange
-        StringBuilder longBody = new StringBuilder(500000);
-        for (int i = 0; i < 25000; i++) {
-            longBody.append(i).append("<dl><dd>");
-        }
-        for (int i = 0; i < 25000; i++) {
-            longBody.append(i).append("</dd></dl>");
-        }
-
-        // Act
-        long start = System.currentTimeMillis();
-        Document doc = Parser.parseBodyFragment(longBody.toString(), "");
-
-        // Assert
-        assertEquals(2, doc.body().childNodeSize());
-        assertEquals(25000, doc.select("dd").size());
-        assertTrue(System.currentTimeMillis() - start < 2000);
-    }
-
     @Test
     public void testInvalidTableContents() throws IOException {
         File in = ParseTest.getFile("/htmltests/table-invalid-elements.html");
